:recycle: Refactor the terms files structure
