+++ grisette.cabal
@@ -98,13 +98,11 @@ library
       Grisette.Internal.Core
       Grisette.Internal.IR.SymPrim
       Grisette.IR.SymPrim
+      Grisette.IR.SymPrim.Data.GeneralFun
       Grisette.IR.SymPrim.Data.IntBitwidth
-      Grisette.IR.SymPrim.Data.Prim.GeneralFun
       Grisette.IR.SymPrim.Data.Prim.Helpers
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.Caches
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermUtils
+      Grisette.IR.SymPrim.Data.Prim.Internal.Caches
+      Grisette.IR.SymPrim.Data.Prim.Internal.Term
       Grisette.IR.SymPrim.Data.Prim.Model
       Grisette.IR.SymPrim.Data.Prim.ModelValue
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Bits
@@ -115,6 +113,9 @@ library
       Grisette.IR.SymPrim.Data.Prim.PartialEval.PartialEval
       Grisette.IR.SymPrim.Data.Prim.PartialEval.TabularFun
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Unfold
+      Grisette.IR.SymPrim.Data.Prim.SomeTerm
+      Grisette.IR.SymPrim.Data.Prim.Term
+      Grisette.IR.SymPrim.Data.Prim.TermUtils
       Grisette.IR.SymPrim.Data.Prim.Utils
       Grisette.IR.SymPrim.Data.SymPrim
       Grisette.IR.SymPrim.Data.TabularFun
