+++ src/Grisette/Core/Data/Class/ToCon.hs
@@ -50,9 +50,9 @@ import Grisette.Core.Data.BV
     WordN (WordN),
   )
 import Grisette.Core.Data.Class.Solvable (Solvable (conView), pattern Con)
+import Grisette.IR.SymPrim.Data.GeneralFun (type (-->))
 import Grisette.IR.SymPrim.Data.IntBitwidth (intBitwidthQ)
-import Grisette.IR.SymPrim.Data.Prim.GeneralFun (type (-->))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( LinkedRep,
     SupportedPrim,
   )
