+++ src/Grisette/Core/Data/Class/ToSym.hs
@@ -58,9 +58,9 @@ import Grisette.Core.Data.BV
     WordN,
   )
 import Grisette.Core.Data.Class.Solvable (Solvable (con))
+import Grisette.IR.SymPrim.Data.GeneralFun (type (-->))
 import Grisette.IR.SymPrim.Data.IntBitwidth (intBitwidthQ)
-import Grisette.IR.SymPrim.Data.Prim.GeneralFun (type (-->))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( LinkedRep,
     SupportedPrim,
   )
