+++ src/Grisette/IR/SymPrim/Data/Prim/Internal/Caches.hs
@@ -6,14 +6,14 @@
 {-# OPTIONS_GHC -fno-cse #-}
 
 -- |
--- Module      :   Grisette.IR.SymPrim.Data.Prim.InternedTerm.Caches
--- Copyright   :   (c) Sirui Lu 2021-2023
+-- Module      :   Grisette.IR.SymPrim.Data.Prim.Internal.Caches
+-- Copyright   :   (c) Sirui Lu 2021-2024
 -- License     :   BSD-3-Clause (see the LICENSE file)
 --
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.IR.SymPrim.Data.Prim.InternedTerm.Caches (typeMemoizedCache) where
+module Grisette.IR.SymPrim.Data.Prim.Internal.Caches (typeMemoizedCache) where
 
 import Control.Concurrent
   ( forkIO,
