+++ src/Grisette/IR/SymPrim/Data/Prim/PartialEval/TabularFun.hs
@@ -17,18 +17,18 @@ module Grisette.IR.SymPrim.Data.Prim.PartialEval.TabularFun
 where
 
 import Grisette.Core.Data.Class.Function (Function ((#)))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( SupportedPrim,
-    Term (ConTerm),
-    conTerm,
-    tabularFunApplyTerm,
-  )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
   ( pevalEqvTerm,
     pevalITETerm,
   )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.PartialEval
   ( totalize2,
+  )
+import Grisette.IR.SymPrim.Data.Prim.Term
+  ( SupportedPrim,
+    Term (ConTerm),
+    conTerm,
+    tabularFunApplyTerm,
   )
 import Grisette.IR.SymPrim.Data.TabularFun
   ( type (=->) (TabularFun),
