+++ src/Grisette/IR/SymPrim/Data/Prim/PartialEval/Unfold.hs
@@ -18,10 +18,6 @@ where
 
 import Control.Monad.Except (MonadError (catchError))
 import Data.Typeable (Typeable)
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( SupportedPrim,
-    Term (ITETerm),
-  )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
   ( pevalITETerm,
   )
@@ -32,6 +28,10 @@ import Grisette.IR.SymPrim.Data.Prim.PartialEval.PartialEval
     TotalRuleUnary,
     totalize,
     totalize2,
+  )
+import Grisette.IR.SymPrim.Data.Prim.Term
+  ( SupportedPrim,
+    Term (ITETerm),
   )
 
 unaryPartialUnfoldOnce ::
