+++ src/Grisette/IR/SymPrim/Data/Prim/SomeTerm.hs
@@ -4,18 +4,18 @@
 {-# LANGUAGE TypeApplications #-}
 
 -- |
--- Module      :   Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm
--- Copyright   :   (c) Sirui Lu 2021-2023
+-- Module      :   Grisette.IR.SymPrim.Data.Prim.SomeTerm
+-- Copyright   :   (c) Sirui Lu 2021-2024
 -- License     :   BSD-3-Clause (see the LICENSE file)
 --
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm (SomeTerm (..)) where
+module Grisette.IR.SymPrim.Data.Prim.SomeTerm (SomeTerm (..)) where
 
 import Data.Hashable (Hashable (hashWithSalt))
 import Data.Typeable (Proxy (Proxy), typeRep)
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( SupportedPrim,
     Term,
     identityWithTypeRep,
@@ -25,10 +25,12 @@ data SomeTerm where
   SomeTerm :: forall a. (SupportedPrim a) => Term a -> SomeTerm
 
 instance Eq SomeTerm where
-  (SomeTerm t1) == (SomeTerm t2) = identityWithTypeRep t1 == identityWithTypeRep t2
+  (SomeTerm t1) == (SomeTerm t2) =
+    identityWithTypeRep t1 == identityWithTypeRep t2
 
 instance Hashable SomeTerm where
   hashWithSalt s (SomeTerm t) = hashWithSalt s $ identityWithTypeRep t
 
 instance Show SomeTerm where
-  show (SomeTerm (t :: Term a)) = "<<" ++ show t ++ " :: " ++ show (typeRep (Proxy @a)) ++ ">>"
+  show (SomeTerm (t :: Term a)) =
+    "<<" ++ show t ++ " :: " ++ show (typeRep (Proxy @a)) ++ ">>"
