+++ src/Grisette/IR/SymPrim/Data/Prim/Term.hs
@@ -1,0 +1,16 @@
+{-# OPTIONS_GHC -Wno-missing-import-lists #-}
+
+-- |
+-- Module      :   Grisette.IR.SymPrim.Data.Prim.Term
+-- Copyright   :   (c) Sirui Lu 2024
+-- License     :   BSD-3-Clause (see the LICENSE file)
+--
+-- Maintainer  :   siruilu@cs.washington.edu
+-- Stability   :   Experimental
+-- Portability :   GHC only
+module Grisette.IR.SymPrim.Data.Prim.Term
+  ( module Grisette.IR.SymPrim.Data.Prim.Internal.Term,
+  )
+where
+
+import Grisette.IR.SymPrim.Data.Prim.Internal.Term
