+++ src/Grisette/IR/SymPrim/Data/Prim/TermUtils.hs
@@ -5,14 +5,14 @@
 {-# LANGUAGE TypeApplications #-}
 
 -- |
--- Module      :   Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermUtils
--- Copyright   :   (c) Sirui Lu 2021-2023
+-- Module      :   Grisette.IR.SymPrim.Data.Prim.TermUtils
+-- Copyright   :   (c) Sirui Lu 2021-2024
 -- License     :   BSD-3-Clause (see the LICENSE file)
 --
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermUtils
+module Grisette.IR.SymPrim.Data.Prim.TermUtils
   ( extractSymbolicsTerm,
     castTerm,
     someTermsSize,
@@ -36,10 +36,10 @@ import Data.Typeable
   ( Typeable,
     cast,
   )
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm
+import Grisette.IR.SymPrim.Data.Prim.SomeTerm
   ( SomeTerm (SomeTerm),
   )
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( SomeTypedSymbol (SomeTypedSymbol),
     SupportedPrim,
     Term
