+++ test/Grisette/Backend/SBV/Data/SMT/TermRewritingTests.hs
@@ -53,7 +53,7 @@ import Grisette.Backend.SBV.Data.SMT.TermRewritingGen
   )
 import Grisette.Core.Data.BV (IntN, WordN)
 import Grisette.Core.Data.Class.Solver (solve)
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( SupportedPrim,
     pformat,
   )
