+++ test/Grisette/Core/Data/Class/EvaluateSymTests.hs
@@ -28,7 +28,7 @@ import Grisette.Core.Data.Class.ModelOps
 import Grisette.Core.Data.Class.SEq (SEq ((.==)))
 import Grisette.Core.Data.Class.Solvable (Solvable (con, isym, ssym))
 import Grisette.Core.Data.Symbol (Symbol (IndexedSymbol))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( TypedSymbol (TypedSymbol),
   )
 import Grisette.IR.SymPrim.Data.Prim.Model
