+++ test/Grisette/Core/Data/Class/SafeSymRotateTests.hs
@@ -22,7 +22,7 @@ import Grisette.Core.Data.Class.SafeSymRotate
   ( SafeSymRotate (safeSymRotateL, safeSymRotateR),
   )
 import Grisette.Core.Data.Class.Solvable (Solvable (con))
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term (LinkedRep)
+import Grisette.IR.SymPrim.Data.Prim.Term (LinkedRep)
 import Grisette.IR.SymPrim.Data.SymPrim (SymIntN, SymWordN)
 import Grisette.Lib.Control.Monad (mrgReturn)
 import Grisette.Lib.Control.Monad.Except (mrgThrowError)
