+++ test/Grisette/Core/Data/Class/TestValues.hs
@@ -14,7 +14,7 @@ import Grisette.Core.Data.Symbol
   ( Identifier,
     Symbol (IndexedSymbol, SimpleSymbol),
   )
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
+import Grisette.IR.SymPrim.Data.Prim.Term
   ( TypedSymbol (TypedSymbol),
   )
 import Grisette.IR.SymPrim.Data.SymPrim (SymBool)
