+++ test/Grisette/IR/SymPrim/Data/Prim/TabularFunTests.hs
@@ -4,18 +4,18 @@
 
 module Grisette.IR.SymPrim.Data.Prim.TabularFunTests (tabularFunTests) where
 
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( Term,
-    conTerm,
-    ssymTerm,
-    tabularFunApplyTerm,
-  )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
   ( pevalEqvTerm,
     pevalITETerm,
   )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.TabularFun
   ( pevalTabularFunApplyTerm,
+  )
+import Grisette.IR.SymPrim.Data.Prim.Term
+  ( Term,
+    conTerm,
+    ssymTerm,
+    tabularFunApplyTerm,
   )
 import Grisette.IR.SymPrim.Data.TabularFun
   ( type (=->) (TabularFun),
